requirements.txt
paho-mqtt
python-dotenv
kafka-python == 2.2.14